# Add non-blocking cycle